# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /users:
    # binds a127 app logic to a route
    x-swagger-router-controller: user.controller.js
    get:
      description: fetch list of users
      # used as the method name of the controller
      operationId: getUserList
      parameters:
        - name: page
          in: query
          description: page no for pagination
          required: false
          type: string
        - name: count
          in: query
          description: no of trains per page
          required: false
          type: string
        - name: select
          in: query
          description: which field to show
          required: false
          type: string
        - name: sort
          in: query
          description: sort query
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /users/admin:
    # binds a127 app logic to a route
    x-swagger-router-controller: userController
    get:
      description: fetch list of admins
      # used as the method name of the controller
      operationId: getAdminList
      parameters:
        - name: page
          in: query
          description: page no for pagination
          required: false
          type: number
        - name: count
          in: query
          description: no of trains per page
          required: false
          type: number
        - name: select
          in: query
          description: which field to show
          required: false
          type: string
        - name: sort
          in: query
          description: sort query
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"          
    post:
      description: fetch list of tickets
      # used as the method name of the controller
      operationId: createAdmin
      parameters:
        - name: data
          in: body
          description: page no for pagination
          required: true
          schema: 
            type: object
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /users/admin/count:
    # binds a127 app logic to a route
    x-swagger-router-controller: userController
    get:
      description: fetch count of admins
      # used as the method name of the controller
      operationId: getAdminCount
      responses:
        "200":
          description: Success
          schema:
            type: object
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 

  ###################### Station APIs ########################
  
  /station:
    # binds a127 app logic to a route
    x-swagger-router-controller: stationController
    get:
      description: fetch list of admins
      # used as the method name of the controller
      operationId: fetchStationList
      parameters:
        - name: page
          in: query
          description: page no for pagination
          required: false
          type: number
        - name: count
          in: query
          description: no of trains per page
          required: false
          type: number
        - name: select
          in: query
          description: which field to show
          required: false
          type: string
        - name: sort
          in: query
          description: sort query
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"          
    post:
      description: fetch list of tickets
      # used as the method name of the controller
      operationId: createStation
      parameters:
        - name: data
          in: body
          description: page no for pagination
          required: true
          schema: 
            type: object
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  ###################### Station APIs End ########################
    
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string